Move FocusNovnc component to CommandAndControl app
