+++ app/packs/src/apps/command_and_control/FocusNovnc.js
